\name{family.cond}
\alias{family.cond}
\title{
  Use family() on a ``cond'' object
}
\description{
  This is a method for the function \code{family()} for objects
  inheriting from class \code{cond}.  See \code{\link[stats]{family}}
  for the general behaviour of this function.}
\usage{
\method{family}{cond}(object, \dots)
}
\arguments{
  \item{object}{
    any object from which a \code{family} object can be extracted.   
  }
  \item{\dots}{
    absorbs any additional argument.  
  }
}
\seealso{
  \code{\link[stats]{family}}
}
\keyword{classes} 
\keyword{methods}
