\name{family.summaryCond}
\alias{family.summaryCond}
\title{
  Use family() on a ``summaryCond'' object
}
\description{
  This is a method for the function \code{family()} for objects 
  inheriting from class \code{summaryCond}.  See 
  \code{\link[stats]{family}} for the general behaviour of this 
  function.
}
\usage{
\method{family}{summaryCond}(object, \dots)
}
\arguments{
  \item{object}{
    any object from which a \code{family} object can be extracted.   
  }
  \item{\dots}{
    absorbs any additional argument.  
  }
}
\seealso{
  \code{\link[stats]{family}}
}
\keyword{classes} 
\keyword{methods}
